LayoutTests/platform/glib/imported/w3c/web-platform-tests/streams/piping/close-propagation-backward.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: releasing lock of reader whose stream is still in readable state
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: releasing lock of reader whose stream is still in readable state
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: releasing lock of reader whose stream is still in readable state
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: releasing lock of reader whose stream is still in readable state
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: releasing lock of reader whose stream is still in readable state
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: releasing lock of reader whose stream is still in readable state
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: releasing lock of reader whose stream is still in readable state

PASS Closing must be propagated backward: starts closed; preventCancel omitted; fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel omitted; rejected cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel = undefined (falsy); fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel = null (falsy); fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel = false (falsy); fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel = 0 (falsy); fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel = -0 (falsy); fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel = NaN (falsy); fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel =  (falsy); fulfilled cancel promise
PASS Closing must be propagated backward: starts closed; preventCancel = true (truthy)
PASS Closing must be propagated backward: starts closed; preventCancel = a (truthy)
PASS Closing must be propagated backward: starts closed; preventCancel = 1 (truthy)
PASS Closing must be propagated backward: starts closed; preventCancel = Symbol() (truthy)
PASS Closing must be propagated backward: starts closed; preventCancel = [object Object] (truthy)
PASS Closing must be propagated backward: starts closed; preventCancel = true, preventAbort = true
PASS Closing must be propagated backward: starts closed; preventCancel = true, preventAbort = true, preventClose = true
